:hammer: Refactor: 파일 구조 추가 변경 #51
