+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/entity/reformOrder/ReformOrderStatus.java
@@ -1,4 +1,4 @@
-package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity;
+package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.reformOrder;
 
 public enum ReformOrderStatus {
     ORDER_COMPLETE, ORDERING // 주문 완료, 주문 중
