+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/entity/reformRequest/ReformRequestStatus.java
@@ -1,4 +1,4 @@
-package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity;
+package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.reformRequest;
 
 public enum ReformRequestStatus {
     REQUEST_WAITING, REQUEST_REJECTED, REQUEST_ACCEPTED // 요청 대기중, 요청 거절, 요청 수락
