+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/repository/EstimateRepository.java
@@ -1,6 +1,6 @@
 package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.repository;
 
-import com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.Estimate;
+import com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.estimate.Estimate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
