+++ src/main/java/com/jjs/ClothingInventorySaleReformPlatform/domain/reform/repository/ProgressRepository.java
@@ -1,6 +1,6 @@
 package com.jjs.ClothingInventorySaleReformPlatform.domain.reform.repository;
 
-import com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.Progressmanagement;
+import com.jjs.ClothingInventorySaleReformPlatform.domain.reform.entity.progressManagement.Progressmanagement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
